Fix bug where Debezium Logical Converters were not used correctly. (#99)
* Fix bug where Debezium Logical Converters were not used correctly.

Also fix some found issues with the debezium date type converter.

Related tests.

* Remove duplicate Kafka Logical converter tests. Condense Logical Converter tests in BigQueryRecordConverterTest to 1 test for Debezium logical types and 1 test for Kafka logical types.

+++ kcbq-connector/src/main/java/com/wepay/kafka/connect/bigquery/convert/logicaltype/DebeziumLogicalConverters.java
@@ -66,7 +66,7 @@ public class DebeziumLogicalConverters {
 
     @Override
     public String convert(Object kafkaConnectObject) {
-      Long daysSinceEpoch = (Long) kafkaConnectObject;
+      Integer daysSinceEpoch = (Integer) kafkaConnectObject;
       long msSinceEpoch = TimeUnit.DAYS.toMillis(daysSinceEpoch);
       java.util.Date date = new java.util.Date(msSinceEpoch);
       return getBQDateFormat().format(date);
